move setup to a script
